refactor(core): Split agent chat bridge modules
