+++ packages/cli/src/modules/agents/integrations/__tests__/slack-integration.test.ts
@@ -1,4 +1,5 @@
 import { SlackIntegration } from '../platforms/slack-integration';
+import type { ChatInstance } from '../chat-integration.service';
 
 describe('SlackIntegration', () => {
 	let integration: SlackIntegration;
@@ -18,6 +19,28 @@ describe('SlackIntegration', () => {
 
 	it('only advertises Slack bot token credentials for agent integrations', () => {
 		expect(integration.credentialTypes).toEqual(['slackApi']);
+	});
+
+	it('extracts the Slack bot user ID for bridge message context', () => {
+		const chat = {
+			getAdapter: jest.fn().mockReturnValue({ botUserId: 'U_BOT' }),
+		} as unknown as ChatInstance;
+
+		expect(integration.getPlatformAgentContext(chat)).toEqual({ agentUserId: 'U_BOT' });
+		expect(chat.getAdapter).toHaveBeenCalledWith('slack');
+	});
+
+	it('strips Slack bot self-mentions before handing text to the agent', () => {
+		expect(integration.prepareInboundText('<@U_BOT> hello', { agentUserId: 'U_BOT' })).toBe(
+			'hello',
+		);
+		expect(integration.prepareInboundText('@U_BOT hello', { agentUserId: 'U_BOT' })).toBe('hello');
+	});
+
+	it('buffers resume responses because Slack action events lack reliable thread context', async () => {
+		await expect(integration.createResumeExecutionContext()).resolves.toEqual({
+			forceBuffered: true,
+		});
 	});
 
 	describe('handleUnauthenticatedWebhook', () => {
